requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
httpx[http2]
vertexai>=0.12.0
pdfplumber>=0.9.0
pillow>=10.0.0
pdf2image>=1.16.0
web3>=6.0.0
ic-py>=1.0.0,<2.0.0
fastapi
python-dotenv
pydantic
google-cloud-aiplatform
git+https://github.com/Vourhey/pinatapy.git@master#egg=pinatapy
numpy
scikit-learn


